Support cloning of ops with regions
This change adds two test cases with a `tensor.pad` who's yielded value
is a value defined by an op outside of the `tensor.pad`. Such paddings
appear in TFLite benchmarks (e.g., MobileNetV3SmallStaticBatch).

`legalizeDispatchWorkgroupOperands` could previously not handle such ops
and left the newly-created dispatch region op in an invalid state. The
TFLite benchmark still passed, but this was just by coincidance. This is
likely due to special handling/folding of constants in the
GreedyPatternRewriter. The second test case replaces the ConstantOp with
a block argument, in which case the GreedyPatternRewriter cannot
accidentally fix the invalid IR; this test case fails without this fix.