Add region_to_workgroup op to FlowExtensions
This transform dialect op converts `dispatch.region` ops into
`dispatch.workgroups` ops. This can eventually replace the part of
DispatchLinalgOnTensors that builds the `workgroups` ops. The op is built
in one step, and there is no need for dropping operands/block arguments
due to tieing etc.

What is still missing a transform that builds the `region` op based
on a heuristic such as the one in DispatchLinalgOnTensors. This heuristic
should put all ops that are needed in the dispatch region into the region.
I.e., this transform should handle both fusion and cloning.