Add transform dialect building blocks for dispatch region formation
This commit adds two new transform dialect ops:
* WrapInDispatchRegionOp: Warp an arbitrary op in a new
  `dispatch.region` op.
* MoveIntoDispatchRegion: Move an op into an existing `dispatch.region`
  op.

These ops will be used to build heuristics for dispatch region
formation. To be able to express the current heuristic in
DispatchLinalgOnTensors.cpp, it is important that both producers and
consumers can be moved (fused) into a `dispatch.region` op.

The implementation of both ops is purpusely inside of static functions,
such the rewrite logic could be called from C++ without using the
transform dialect, if necessary. (Nevertheless, the transform dialect
makes it easy to test the functionality indepently.)